checkpoint, some scaffolding done in move gen, init moving pieces is done

+++ src/board/display.rs
@@ -11,9 +11,9 @@ impl fmt::Display for Board {
             for (piece, bitboard) in bitboards.iter().enumerate() {
                 for file in (Files::A..=Files::H).rev() {
                     for rank in (Ranks::R1..=Ranks::R8).rev() {
-                        if (bitboard.bits() >> ((rank * 8) + file) as usize) & 1 != 0 {
+                        if (bitboard.bits() >> (rank * 8) + file) & 1 != 0 {
                             let piece_str = format!("{}", Piece::new(piece));
-                            board[rank as usize][file as usize] = match side {
+                            board[rank][file] = match side {
                                 Sides::WHITE => piece_str.chars().next().unwrap(),
                                 Sides::BLACK => {
                                     piece_str.chars().next().unwrap().to_ascii_lowercase()
